Two sensing channels, calculates angle now, evt driven (non-blocking)

+++ include/mpos.h
@@ -11,6 +11,7 @@
 
 #include <stdint.h>
 #include <math.h>
+#include <inttypes.h>
 #include "nrfx_saadc.h"
 #include "app_error.h"
 
@@ -26,7 +27,7 @@ int16_t mpos_test_convert(void);
 
 void mpos_test_convert_event_activate(void);
 
-float angle(int16_t hall_0, int16_t hall_1);
+double angle(int16_t hall_0, int16_t hall_1);
 
 void display_value(void);
